did some code cleaning
